pyproject.toml
[project]
name = "codec-convert"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
 "colored>=2.3.1",
 "ffmpeg-python>=0.2.0",
]

# [project.scripts]
# hello = "hello_world:hello"
# [project.scripts]
# codec_convert = "codec_convert:main"
[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
